Changed from int to float
I did that so decimals work as they do not in int so for example 3.14 works now

+++ main.cpp
@@ -2,8 +2,8 @@
 
 
 int main() {
-    int first_number;
-    int second_number;
+    float first_number;
+    float second_number;
     char calc;
     std::cout << "Please enter your first number! \n :";
     std::cin >> first_number;
